Joined Inventory Data and Display
Created inventory script to handle the displaying of the inventory to
the window when the game is paused.  Items are now stored in that script
and can be manipulated from the player script.  Removing items is added,
but has not been tested because we don't have any reason to remove them
yet.