Inventory Work - "Finishing Touches"
Inventory is more or less finished, but there will probably be bugs down
the road as we start to implement it.  Here are the changes:

Fixed some state change issues in GameManager, so the game knows what
state it was in before pausing and can go back to that state.
The player will stop moving when they start viewing an object, to
prevent them sliding out of view if they are running at it.
When using an object on another one, the player can right-click to put
the object back in their inventory.
The item script now has a "reagent ID" which is an int of the itemID
that will activate it (i.e. - Door's reagent ID will be the item ID of
the matching key).  When used, the item will be removed if it is
successfully used, or the player will remain in Object use mode if the
items don't match.  Right now, only debug statements print, but
on-screen text can be changed in the future.

Future work: On screen text is getting messy and should be moved to its
own script.  Right now, I have a method that pushes a global string to
the UI every frame, but I feel as though it should be its own component.
Also, we can look at inheriting from the standard item script for
specific items that do specific actions.