Added Main Menu, Scene Loader, Changed Prefabs
Added a main menu scene that should be the launch scene whenever we
build standalone.  This currently has 4 buttons, to load the 3 demo
scenes and one to get to the options screen (currently blank).

Updated the GameManager and Player to be prefabs in the foyer scene.
**IMPORTANT: When making changes to the scripts make sure to hit "apply"
to the prefabs so the changes persist across scenes.

Lastly, ensure that objects do not have 2 colliders on them, as the
object viewer will only turn 1 off and the second collider will push the
player around when viewing.