Better Highlighting, Object View Distance changed to variable
You now NEED to define the starting object interaction distance in the
Item script (inspector) for each item.  Zooming will soon be removed, so
place it in a location such that it will not clip through the player,
even when rotated.

Highlighting raycasting is now fixed, and will more accurately highlight
when over the crosshair.