Moved examine text handling to texthandler class, better notifications
New methods added to texthandler for starting a timer, displaying
examine text in the examine node, clearing text out of that node.

Picking up an object now displays that it has been added to your
inventory.  When item is selected from inventory, highlight text changes
to "use item on x"