Added Animatic option to Dialogue
You can now identify a line in your event dialogue with "a" and then the
text, which will cause the corresponding animatic to be played at that
time.  It's a really janky system as it sits now, but I'm thinking about
reworking how events are handled in the near future.