Added mouse hover to Inventory, changed lighting of viewed objects
Added a description field to the item script where you can add your text
description of the item (what you want it to say when hovered over in
the inventory)

Hovering over an inventory item now displays the name and information in
a box.  (still need to adjust the box position based on which slot is
being hovered over)

Changed the shader when the object is being viewed to Unlit Texture -
this will help annoying reflections when the object is being looked at
and moved around but will make some objects look unnatural (vase, for
example).