UI hover box update and applied all prefabs
Hooked up the UI hover text to move next to the slot that you are
examining.  Moved the text box to the bottom of the hierarchy so that it
is always drawn last (over top of everything else).

Additionally, applied all prefabs in the demo scene so that the changes
carry over to other scenes.