Added rest of dialogue events, moved mp3s to resources folder, disabled stair sliding, colliders in office
Added the rest of the events that we have the mp3 recordings for.  The
Excel sheet on Drive has also been updated with the list of events.

Set Player rigidbody to kinematic when viewing objects, so it will no
longer slide down the stairs.

Added colliders to the office objects and fixed bouncing with pictures
and beer cans.

Moved all audio files to resources folder so that they can be loaded in
real time when the scene is built to executable.